requirements.txt
openai>=1.35.7
streamlit>=1.36.0
python-dotenv>=1.0.1
firebase-admin>=6.5.0
scikit-learn>=1.5.0
numpy>=1.24.4
joblib>=1.3.2
google-cloud-firestore>=2.16.0
google-auth>=2.30.0
grpcio>=1.60.0
pyyaml>=6.0.0
